renamed FilterHelpers trait and moved it into the FilterHelper object
