+++ geomesa-core/src/main/scala/org/locationtech/geomesa/core/index/FilterHelper.scala
@@ -28,7 +28,7 @@ import org.opengis.feature.simple.SimpleFeatureType
 import org.opengis.filter._
 import org.opengis.filter.expression.{Expression, Literal, PropertyName}
 import org.opengis.filter.spatial._
-import org.opengis.filter.temporal.{BinaryTemporalOperator, During, Before, After}
+import org.opengis.filter.temporal.{After, Before, During}
 import org.opengis.temporal.Period
 
 import scala.collection.JavaConversions._
@@ -211,4 +211,37 @@ object FilterHelper {
       case f: Filter => Seq(f)
     }
   }
+}
+
+trait IndexFilterHelpers {
+
+  def buildFilter(geom: Geometry, interval: Interval): KeyPlanningFilter =
+    (IndexSchema.somewhere(geom), IndexSchema.somewhen(interval)) match {
+      case (None, None)       =>    AcceptEverythingFilter
+      case (None, Some(i))    =>
+        if (i.getStart == i.getEnd) DateFilter(i.getStart)
+        else                        DateRangeFilter(i.getStart, i.getEnd)
+      case (Some(p), None)    =>    SpatialFilter(p)
+      case (Some(p), Some(i)) =>
+        if (i.getStart == i.getEnd) SpatialDateFilter(p, i.getStart)
+        else                        SpatialDateRangeFilter(p, i.getStart, i.getEnd)
+    }
+
+  def netGeom(geom: Geometry): Geometry =
+    Option(geom).map(_.intersection(IndexSchema.everywhere)).orNull
+
+  def netInterval(interval: Interval): Interval = interval match {
+    case null => null
+    case _    => IndexSchema.everywhen.overlap(interval)
+  }
+
+  def netPolygon(poly: Polygon): Polygon = poly match {
+    case null => null
+    case p if p.covers(IndexSchema.everywhere) =>
+      IndexSchema.everywhere
+    case p if IndexSchema.everywhere.covers(p) => p
+    case _ => poly.intersection(IndexSchema.everywhere).
+      asInstanceOf[Polygon]
+  }
+
 }
